+++ torch_frame/__init__.py
@@ -1,4 +1,5 @@
-from .stype import stype, numerical, categorical, text_embedded
+from .stype import (stype, numerical, categorical, text_embedded,
+                    multi_categorical)
 from .data import TensorFrame
 from .typing import TaskType, DataFrame, NAStrategy
 from torch_frame.utils import save, load, cat  # noqa
@@ -15,6 +16,7 @@ __all__ = [
     'numerical',
     'categorical',
     'text_embedded',
+    'multi_categorical',
     'TaskType',
     'NAStrategy',
     'TensorFrame',