+++ CHANGELOG.md
@@ -3,11 +3,27 @@
 All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
+
+## [Unreleased]
+
+### Added
+
+- Added `multi_categorical` stype ([#128](https://github.com/pyg-team/pytorch-frame/pull/128))
+- Added `MultiNestedTensor` ([#149](https://github.com/pyg-team/pytorch-frame/pull/149))
+
+### Changed
+
+### Deprecated
+
+### Removed
+
+### Fixed
+
+
 ## [0.1.0] - 2023-10-23
 
 ### Added
 
-- Added `MultiNestedTensor` ([#149](https://github.com/pyg-team/pytorch-frame/pull/149/files))
 - Added `Mercari` dataset ([#123](https://github.com/pyg-team/pytorch-frame/pull/123/files))
 - Added the model performance benchmark script ([#114](https://github.com/pyg-team/pytorch-frame/pull/114))
 - Added `DataFrameBenchmark` ([#107](https://github.com/pyg-team/pytorch-frame/pull/107))